cleanup wizard structure so it really works
