+++ src/features/colinks/WizardProgress.ts
@@ -1,0 +1,9 @@
+export type WizardProgress = {
+  address: string;
+  onCorrectChain: boolean;
+  hasName: boolean;
+  hasRep: boolean;
+  hasCoSoul: boolean;
+  hasOwnKey: boolean;
+  hasOtherKey: boolean;
+};
