+++ src/features/cosoul/CoLinksMintPage.tsx
@@ -7,7 +7,6 @@ import { SingleColumnLayout } from 'ui/layouts';
 import { CoSoulArt } from './art/CoSoulArt';
 import { CoLinksCoSoulArtContainer } from './CoLinksCoSoulArtContainer';
 import { coSoulCloud } from './CoSoulArtContainer';
-import { CoSoulDetails } from './CoSoulDetails';
 import { getCoSoulData, QUERY_KEY_COSOUL_PAGE } from './getCoSoulData';
 
 export const artWidthMobile = '320px';
@@ -48,7 +47,6 @@ export const CoLinksMintPage = ({ minted }: { minted: boolean }) => {
             <CoSoulArt pGive={cosoul_data.totalPgive} address={address} />
             <Box css={{ ...coSoulCloud, zIndex: -1 }} />
           </CoLinksCoSoulArtContainer>
-          <CoSoulDetails cosoul_data={cosoul_data} minted={minted} />
         </SingleColumnLayout>
       )}
     </>
